Refactor duplicate interfaces.
