+++ frontend/src/components/Reviews.tsx
@@ -9,7 +9,7 @@ import {
     Typography,
 } from '@material-ui/core';
 import Rating from '@material-ui/lab/Rating';
-import { Feedback } from '../store/meetings/actions';
+import { Feedback } from '../store/interfaces';
 import { CardImage, Middle } from './meetings/MeetingAdd';
 import Blockies from 'react-blockies';
 import EtherService from '../services/EtherService';
