+++ frontend/src/containers/meetings/MeetingAdd.tsx
@@ -3,9 +3,9 @@ import { Dispatch } from 'react';
 import { compose, Action } from 'redux';
 import { connect } from 'react-redux';
 import { History } from 'history';
-import { actions as clubActions, Club } from '../../store/clubs/actions';
-import { actions as meetingActions, Meeting, GroupHashAndAddress } from '../../store/meetings/actions';
-import { User } from '../../store/users/actions';
+import { actions as clubActions } from '../../store/clubs/actions';
+import { actions as meetingActions } from '../../store/meetings/actions';
+import { User, Club, Meeting, GroupHashAndAddress } from '../../store/interfaces';
 import { actions as loadingActions } from '../../store/loading/actions';
 import { IAppState } from '../../store/index';
 import { MeetingAdd as Component } from '../../components/meetings/MeetingAdd';
