+++ frontend/src/store/loading/actions.ts
@@ -1,18 +1,6 @@
 import actionCreatorFactory from 'typescript-fsa';
 import { AppActions } from '../constants';
 
-export interface Loading {
-  cachedMeeting: boolean;
-  meetingDeployment: boolean;
-  rsvpConfirmation: boolean;
-  rsvpCancellationConfirmation: boolean;
-  startMeetingConfirmation: boolean;
-  endMeetingConfirmation: boolean;
-  cancelMeetingConfirmation: boolean;
-
-  markAttendanceConfirmation: boolean;
-  markAbsenceConfirmation: boolean;
-}
 
 const actionCreator = actionCreatorFactory();
 
@@ -25,7 +13,6 @@ export const UpdateEndMeetingConfirmationLoading = actionCreator<boolean>(AppAct
 export const UpdateCancelMeetingConfirmationLoading = actionCreator<boolean>(AppActions.UPDATE_CANCEL_MEETING_CONFIRMATION_LOADING);
 export const UpdateMarkAttendanceConfirmationLoading = actionCreator<boolean>(AppActions.UPDATE_MARK_ATTENDANCE_CONFIRMATION_LOADING);
 export const UpdateMarkAbsenceConfirmationLoading = actionCreator<boolean>(AppActions.UPDATE_MARK_ABSENCE_CONFIRMATION_LOADING);
-
 
 export const actions = {
   UpdateCachedMeetingLoading,
