+++ frontend/src/store/users/actions.ts
@@ -1,31 +1,7 @@
 import actionCreatorFactory from 'typescript-fsa';
 import { AppActions } from '../constants';
-import { Feedback } from '../meetings/actions';
+import { Feedback, User } from '../interfaces';
 
-// TODO refactor duplicate
-export const ModelType = {
-  USER: 'user',
-  PENDING: 'pending',
-  MEETING: 'meeting',
-  CLUB: 'club',
-};
-
-export interface User {
-  _id: string;
-  type: string;
-  admins: ReadonlyArray<string>;
-
-  data: {
-    ensDomain: string,
-    feedback: ReadonlyArray<Feedback>,
-
-    // list of meeting wallets (smart contract address) linked to this user profile per status
-    cancel: ReadonlyArray<string>;
-    rsvp: ReadonlyArray<string>;
-    attend: ReadonlyArray<string>;
-    withdraw: ReadonlyArray<string>;
-  };
-}
 
 const actionCreator = actionCreatorFactory();
